refactor: Embedding, LLM API 제거 및 util로 이동
