+++ src/main/java/com/wellmeet/WellMeet_Recommendation/exception/GlobalExceptionHandler.java
@@ -1,20 +1,19 @@
 package com.wellmeet.WellMeet_Recommendation.exception;
 
 import jakarta.servlet.http.HttpServletRequest;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.slf4j.Logger;
 
 @RestControllerAdvice
+@Slf4j
 public class GlobalExceptionHandler {
 
-    private static final Logger log = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-
     @ExceptionHandler(WellMeetException.class)
-    public ResponseEntity<ErrorResponse> handleOAuthClientException(WellMeetException exception) {
+    public ResponseEntity<ErrorResponse> handleWellMeetException(WellMeetException exception) {
         return toResponse(exception.getStatus(), exception.getMessage());
     }
