+++ src/main/java/com/wellmeet/WellMeet_Recommendation/restaurant/controller/RestaurantController.java
@@ -1,6 +1,5 @@
 package com.wellmeet.WellMeet_Recommendation.restaurant.controller;
 
-import com.wellmeet.WellMeet_Recommendation.restaurant.dto.RecommendRestaurantRequest;
 import com.wellmeet.WellMeet_Recommendation.restaurant.dto.RestaurantCreateRequest;
 import com.wellmeet.WellMeet_Recommendation.restaurant.dto.RestaurantResponse;
 import com.wellmeet.WellMeet_Recommendation.restaurant.service.RestaurantService;
@@ -10,7 +9,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
 import java.util.List;
 
 @RestController
@@ -19,15 +17,15 @@ public class RestaurantController {
 
     private final RestaurantService restaurantService;
 
-    @PostMapping("/api/restaurants")
+    @PostMapping("/api/restaurant")
     @ResponseStatus(HttpStatus.CREATED)
     public RestaurantResponse createRestaurant(@RequestBody RestaurantCreateRequest request) {
         return restaurantService.saveRestaurant(request);
     }
 
-    @PostMapping("/api/restaurants/recommend")
+    @PostMapping("/api/restaurant/recommend")
     @ResponseStatus(HttpStatus.OK)
-    public List<RestaurantResponse> recommendRestaurants(@RequestBody RecommendRestaurantRequest request) {
-        return restaurantService.recommendRestaurant(request.getQuery());
+    public List<RestaurantResponse> recommendRestaurants(@RequestBody String query) {
+        return restaurantService.recommendRestaurant(query);
     }
 }
