--- src/main/java/com/wellmeet/WellMeet_Recommendation/restaurant/dto/RecommendRestaurantRequest.java
@@ -1,14 +1,0 @@
-package com.wellmeet.WellMeet_Recommendation.restaurant.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class RecommendRestaurantRequest {
-    private String query;
-
-    public RecommendRestaurantRequest(String query) {
-        this.query = query;
-    }
-}
